update: publish to jetstream after rustbot finishes talking
Previoously we would publish to jetstream as soon as
write_stream in TTS task completes. The issue with that is that since we
are playing audio in an async task, we never wait for the actual
audio to finish. We try to do that here. We also handle
audio player exit better here.

Signed-off-by: Milos Gajdos <milosthegajdos@gmail.com>